+++ components/homePage/Banner.tsx
@@ -4,8 +4,6 @@ import {
   Button,
   Grid,
   Typography,
-  Modal,
-  DialogContent,
 } from "@mui/material";
 import Image from "next/image";
 import mainImage from "../../public/images/homePage/heroSectionImage/mainImage.webp";
@@ -14,7 +12,7 @@ import ArrowRightIcon from "@mui/icons-material/ArrowRight";
 import { bannerDetails } from "../../data/homePage";
 import { lora } from "../../app/fonts";
 import * as React from "react";
-import modalImage from "../../public/images/homePage/dialogBox/thumbUp.webp";
+
 import { useRouter } from "next/navigation";
 
 import {
@@ -25,29 +23,11 @@ import {
   buttonTextStyle,
 } from "./styles";
 
-// modal styles
-const style = {
-  position: "absolute" as "absolute",
-  top: "50%",
-  left: "50%",
-  transform: "translate(-50%, -50%)",
-  width: { xs: "85%", sm: "60%", md: "35%" },
-  height: "auto",
-  bgcolor: "background.paper",
-  boxShadow: 50,
-  p: 4,
-  display: "flex",
-  flexDirection: "column",
-  alignItems: "center",
-  justifyContent: "center",
-  borderRadius: "10px",
-};
+
 
 /* Banner functional component */
 const Banner = () => {
-  const [open, setOpen] = React.useState(false);
-  const handleOpen = () => setOpen(true);
-  const handleClose = () => setOpen(false);
+
   const router = useRouter();
 
 
@@ -139,95 +119,7 @@ const Banner = () => {
               </Box>
               {bannerDetails.bannercta2Caption}
             </Button>
-            {/* modal display the success */}
-            <Modal
-              keepMounted
-              open={open}
-              onClose={handleClose}
-              aria-labelledby="keep-mounted-modal-title"
-              aria-describedby="keep-mounted-modal-description"
-            >
-              <Box sx={style}>
-                <DialogContent
-                  dividers
-                  style={{
-                    display: "flex",
-                    flexDirection: "column",
-                    alignItems: "center",
-                    justifyContent: "center",
-                    backgroundColor: "lightgreen",
-                    borderRadius: "10px",
-                  }}
-                >
-                  {/* modal image */}
-                  <Image
-                    src={modalImage}
-                    alt="test"
-                    style={{ maxWidth: "25%", height: "auto" }}
-                  />
-                </DialogContent>
-
-                <DialogContent
-                  dividers
-                  style={{
-                    display: "flex",
-                    flexDirection: "column",
-                    alignItems: "center",
-                    justifyContent: "center",
-                    textAlign: "center",
-                  }}
-                >
-                  {/* modal title */}
-                  <Typography
-                    id="keep-mounted-modal-title"
-                    variant="h6"
-                    component="h2"
-                    sx={{
-                      color: "#54BC2F",
-                      fontSize: { xs: "15px", sm: "20px", md: "22px" },
-                    }}
-                  >
-                    {bannerDetails.modalTitle}
-                  </Typography>
-                  {/* modal text */}
-                  <Typography
-                    id="keep-mounted-modal-description"
-                    sx={{
-                      mt: 2,
-                      fontSize: { xs: "12px", sm: "18px", md: "20px" },
-                    }}
-                  >
-                    {bannerDetails.modalContentText}
-                  </Typography>
-                  {/* modal button */}
-                  <Button
-                    variant="contained"
-                    color="success"
-                    autoFocus
-                    onClick={handleClose}
-                    sx={{
-                      backgroundColor: "#52AA44",
-                      width: "50%",
-                      height: "5%",
-                      borderRadius: "58px",
-                      display: "flex",
-                      alignItems: "center",
-                      justifyContent: "center",
-                      marginTop: "20px",
-                    }}
-                  >
-                    <Typography
-                      sx={{
-                        color: "#FFFFFF",
-                        fontSize: { xs: "12px", sm: "18px", md: "15px" },
-                      }}
-                    >
-                      {bannerDetails.captionButton}
-                    </Typography>
-                  </Button>
-                </DialogContent>
-              </Box>
-            </Modal>
+            
           </Box>
         </Grid>
         {/* Right side of the banner - image */}
